fix(charts): geth flags (#1957)
## Summary
changes to how flags condition is evaluated and used.
## Background
Our current evm-rollup charts has a bug, the condition which helps
configuring geth flags is not being evaluated and always acts as `true`.
This PR changes the condition templating.
## Changes
- conditions are now being templated based on the key value. Also makes
sure this value is `true` if the condition exists.

## Testing
- locally running smoke test against a cluster 
- using the helm command for templating: ` helm template my-chart
./charts/evm-stack -f ./dev/values/rollup/dev.yaml`
## Changelogs
No updates required.