fix(sequencer)!: fix fee storage enum ordering (#1983)
## Summary
Reverts a state breaking change in fee values.

## Background
We added a fee variable to the middle of an enum resulting in Borsch
serialization breaking for existing stored state.

## Changes
- Move new enum to the end of the list

## Testing
Tests to be in follow up PR ensuring we don't make such breaking changes
in the future.

## Changelogs
I did not make a changelog entry here since we haven't cut a release
since the breaking behaivior was added.