feat(proto): create v2 execution API (#1962)
## Summary
Creates v2 execution API to support upcoming v2 changes to Conductor and
Geth.

## Background
These changes are meant to support the upcoming v2 upgrades to
[Conductor](https://github.com/astriaorg/astria/pull/1928) and
[astria-geth](https://github.com/astriaorg/astria-geth/pull/59), which
will facilitate the ability to have rollup forks, driven by the need to
migrate Forma to Mainnet.

## Changes from `v1`
- Removed `GenesisInfo` and associated RPC.
- Removed `BatchGetBlocks` RPC and associated types.
- Removed `GetCommitmentState` RPC.
- Introduced **`ExecutionSession`** and associated RPC
`CreateExecutionSession`:
- An `ExecutionSession` is the execution client-side's current session,
which starts at a specific block number and can end at one as well.
- Upon startup and after each restart (if stop numbers are specified),
`CreateExecutionSession` should be called, to which the server should
respond with the current session, containing an ID for the session, the
`CommitmentState`, and `ExecutionSessionParams`.
- Added `ExecutionSessionParams`, which contains the necessary
configuration for the execution client to run for the entirety of the
session, consisting of the following:
     - `rollup_id`
- `rollup_start_block_number`, the first block to be executed against
the rollup
- `rollup_end_block_number`, the final block to be executed against the
rollup before ending the current execution session.
     - `sequencer_chain_id`
- `sequencer_start_block_height`, the sequencer height to map to
`rollup_start_block_number`.
     - `celestia_chain_id`
- `celestia_max_search_height_look_ahead` (formerly
`celestia_block_variance`)
- Added `session_id` to `ExecuteBlockRequest` and
`UpdateCommitmentStateRequest`, so that the server can validate the
calls are for the correct execution session.
- Changed `Block` to `ExecutedBlockMetadata` and made corresponding
service changes.
- Added `ExecuteBlockResponse` to follow AIP.
- Changed `CommitmentState.base_celestia_height` to
`lowest_celestia_search_height`
- Represented all hashes as Strings so that they are displayed in
whichever format the hashes are in, as opposed to Base64.
- Separated all top-level proto entities into separate files (per [1-1-1
best practice](https://protobuf.dev/best-practices/1-1-1/))

## Testing
No testing needed.

## Changelogs
Protos have no changelog

---------

Co-authored-by: Jordan Oroshiba <jordan@astria.org>
Co-authored-by: Superfluffy <janis@astria.org>