feat(sequencer)!: implement `RecoverClient` action  (#2008)
## Summary
implement `RecoverClient` action which replaces an expired or otherwise
inactive IBC client.

this is based off `MsgRecoverClient` implemented in ibc-go v8:
https://github.com/cosmos/ibc-go/blob/a71577cc1bf242f8c99360777106b1321503b17f/proto/ibc/core/client/v1/tx.proto#L126

## Background
there is no way to otherwise replace an expired IBC client without
creating a new one and creating a governance proposal on the
counterparty chain to change the client there, which is a lot more
hassle than just being able to replace an expired client on one chain.

## Changes
- implement `RecoverClient` action 
- the implementation is based off the following ibc-go implementation: 
-
https://github.com/cosmos/ibc-go/blob/a71577cc1bf242f8c99360777106b1321503b17f/modules/core/02-client/keeper/client.go#L121
-
https://github.com/cosmos/ibc-go/blob/a71577cc1bf242f8c99360777106b1321503b17f/modules/light-clients/07-tendermint/light_client_module.go#L179
-
https://github.com/cosmos/ibc-go/blob/a71577cc1bf242f8c99360777106b1321503b17f/modules/light-clients/07-tendermint/proposal_handle.go#L29
- i had to export some types from penumbra (ConsensusStateWriteExt,
ClientState) which is why it's on a different commit (commit is based
off v0.80.7)

## Testing
not yet, will be done on devnet

## Changelogs
changelogs updated

## Breaking Changelist
- a new action is added, which is a breaking change.

## Related Issues
closes https://github.com/astriaorg/astria/issues/2005