feat(sequencer, core, conductor)!: merge upgrades/price-feed feature branch to main (#2085)
## Summary
This merges the long-running feature branch `feat/oracle` to `main`.

## Background
This is in preparation for deploying the new functionality provided in
the feature branch, i.e. sequencer upgrades and price-feed data.

## Changes
- This PR is only a merge of the feature branch, comprising many
individual PRs to the feature branch along with PRs merging changes from
main into the feature branch to keep it in sync.

## Testing
The new features are well covered in new unit and system tests.

## Changelogs
Changelogs already updated in the feature branch.

## Metrics
New metrics:
- `astria_sequencer_extended_commit_info_bytes`
- `astria_sequencer_extend_vote_duration_seconds`
- `astria_sequencer_extend_vote_failure_count`
- `astria_sequencer_verify_vote_extension_failure_count`

## Breaking Changelist
- The addition of price feed data via vote extensions and support for
sequencer upgrades are the two major breaking changes. However, these
changes will be applied once the sequencer networks upgrade at given
future block heights. Sequencers running the code in this PR will be
compatible with existing networks until the upgrades are activated.

---------

Co-authored-by: noot <36753753+noot@users.noreply.github.com>
Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>
Co-authored-by: elizabeth <elizabethjbinks@gmail.com>
Co-authored-by: Itamar Reif <9663129+itamarreif@users.noreply.github.com>
Co-authored-by: quasystaty <ido@astria.org>
Co-authored-by: Jordan Oroshiba <jordan@astria.org>
Co-authored-by: Ethan Oroshiba <ethan@astria.org>
Co-authored-by: Josh Bowen <josh@astria.org>
Co-authored-by: Amrik Ajimal <aajimal@users.noreply.github.com>
Co-authored-by: Jesse Snyder <jessetsnyder@gmail.com>