feat(chart): split up config variables in rollup chart, add default genesis files (#2046)
## Summary
Updates the rollup chart to support including genesis based on raw file,
and moves things out of a nested config.

This is also updated for the newest version of flame devnet, and the
chart is currently being used for running it.

## Background
In order to support the evm-cluster app it was useful to restructure
some of the configs. Additionally wanted to make it so that changes to
genesis config setup wouldn't affect production nodes, and make it
easier for a third party to run a node if they desired to.

## Changes
- Default values for images and genesis information based on the network
provided.
- removed `config` in values and flattened to `geth` and `conductor`
entries.
- updated all flags to me a map instead of a list. This enables adding a
single new entry or overriding a single value without needing to copy
all defaults.
- added ability to specify specific module log levels for geth via
values
- migrated to using the geth standard of number for logging verbosity.
- allow full specification of conductor logging levels.
- moved celestia node values in `conductor` configuration.

## Testing
Chart was used for the deployment of new evm for dusk-11, CI/CD smoke
test.

## Changelogs
No updates required.

## Breaking Changelist
- This is breaking on values files used for deployments vs previous
charts. Major updates made to configuration where some pieces are
flattened and others moved around. See the diff of evm-rollup values
file for more details.