chore(sequencer): add app-level mempool tests for transaction inclusion (#2165)
## Summary
Added app-level mempool tests to ensure that transactions are
successfully removed from the mempool with an "included" status in
regular execution.

## Background
#2133 added support for `TransactionStatus`es, namely creating a new
`RemovalReason::IncludedInBlock`. Tests were added to ensure that
transaction IDs which were passed to the mempool in `App::commit` were
successfully marked as "included", but no tests were added at a higher
level to ensure that transactions are successfully conveyed between the
different app methods and then to mempool. This led to a silent failure
when the PR was first drafted, and the tests were deemed best suited for
a followup.

These changes aim to address the three possible execution flows within
the app to ensure executed transaction IDs are successfully passed to
the mempool in `commit` no matter when they were executed:
1. `prepare_proposal` (execution here) -> `process_proposal` ->
`finalize_block` -> `commit`
2. `process_proposal` (execution here) -> `finalize_block` -> `commit`
3. `finalize_block` (execution here) -> `commit`

## Changes/Testing
- Added tests for the 3 flows listed above in `app/tests/mempool`.
- Removed an eyre hook installation that was causing failure.

## Changelogs
No updates required.

## Related Issues
closes #2146