feat(sequencer): add tx result to "executed" transaction status response (#2159)
## Summary
Adds the `ExecTxResult` to `TransactionStatus::Executed` for the given
transaction.

## Background
Previously, the `TransactionStatus::Executed` only contained the height
at which the transaction was executed. It would be good UX to included
the results of the transaction as well, which this change aims to
address.

## Changes
- Adds `ExecTxResult`, `Event`, and `EventAttribute` to vendored protos,
and added `result` field to `TransactionStatus::Executed`.
- Implements `Protobuf` for the Tendermint Rs counterparts to the new
generated types.
- Rolls transaction IDs into the `tx_results` collections throughout
`sequencer::App`, and includes the collection in `WriteBatch` to be
passed into `run_maintenance`.

## Testing
Added the result to existing test for `TransactionStatus::Executed`.

## Changelogs
Changelog updated.

## Related Issues
closes #2153