chore(ci): rewrite smoke test in python (#2141)
## Summary
Migrates smoke test to python in a similar manner to the Aspen upgrade
test.

## Background
Currently, we heavily rely on `just` for running our end-to-end tests, a
tool which is intended to be used as a command runner. By migrating our
e2e tests to Python, we improve readability and maintainability and
decrease our reliance on a tool which is not meant for this purpose.

## Changes
- Added `smoke_test` to system tests, making accordant changes to
`evm-rollup-restart-test` as well.
- Renamed `evm-rollup-restart-test` to `evm-restart-test` for brevity.
- Extended helpers as necessary.
- Added `defaults` to system test helpers.
- Removed unnecessary helpers in `charts/just` for running smoke tests.

## Testing
Passing test.

## Changelogs
No updates required.

## Related Issues
closes #2067